% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{build_model}
\alias{build_model}
\title{Run a fit on a given dataset}
\usage{
build_model(
  data,
  c0 = 15,
  history_window = 200,
  penalty_factor = 5,
  rss_bias = 30,
  start = 10,
  l1_bounds = seq(from = 200, to = 0, by = -2),
  lag_bounds = -1,
  date_column = "day|date",
  response_column = "plt_used",
  show_progress = TRUE,
  plot_losses = TRUE
)
}
\arguments{
\item{data}{the dataset}

\item{c0}{the lower bound on remaining "fresh" inventory (used in optimization)}

\item{history_window}{number of days to look back}

\item{penalty_factor}{penalty for shortage specified by doctors}

\item{rss_bias}{amount by which we should bias the RSS calculation upward for the loss}

\item{start}{the first day in the dataset that the model's predictions are evaluated}

\item{l1_bounds}{vector containing the possible values of the l1 bound on coefs aside from
days of the week and seven day moving average}

\item{lag_bounds}{vector containing possible values of the bound on the seven day moving average
(lag) coefficient}

\item{date_column}{the name of the date or day number column as a
regex, default is "day|date" i.e. day or date}

\item{response_column}{the name of the response column, default is
"plt_used"}

\item{show_progress}{a TRUE/FALSE flag for showing progress,
default TRUE}

\item{plot_losses}{a TRUE/FALSE flag for whether we plot loss values by hyperparameter
(helpful for model diagnostics). Default true.}
}
\description{
The dataset should have one more column than the predictors and
response. This additional column, which is typically a date or day
number, has to be named "date" or "day"
}
\examples{

data_file <- system.file("extdata", "sample_data.Rdata", package = "pip")
load(data_file)

# runs single_lpSolve to produce the model with the lowest CV loss using
# hyperparameters l1_bounds and lag_bounds (LASSO penalty)
model <- build_model(lpdata, c0 = 15, history_window = 100, penalty_factor = 5,
                     rss_bias = 30, l1_bounds = l1_bounds, lag_bounds = lag_bounds)

# generate single total platelet usage prediction for the next 3 days.
pred <- predict_three_day_sum(model, newdata)

}
