% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{compute_cv_loss}
\alias{compute_cv_loss}
\title{Compute the Cross-Validation Loss for each hyperparameter}
\usage{
compute_cv_loss(w_cv, r_cv, s_cv, n, lo_inv_limit, hi_inv_limit)
}
\arguments{
\item{w_cv}{a vector of waste determined by pip::compute_prediction_statistics}

\item{r_cv}{a vector of remaining inventory determined by pip::compute_prediction_statistics}

\item{s_cv}{a vector of inventory shortage as determined by pip::compute_prediction_statistics}

\item{n}{number of days to consider (length of w_cv, s_cv)}

\item{lo_inv_limit}{threshold for insufficient inventory below which to penalize (shortage risk)}

\item{hi_inv_limit}{threshold for surfeit inventory above which to penalize (wastage risk)}
}
\description{
Compute the Cross-Validation Loss for each hyperparameter
}
