% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{compute_prediction_statistics}
\alias{compute_prediction_statistics}
\title{Compute the prediction statistics}
\usage{
compute_prediction_statistics(y, t_pred, initial_expiry_data = c(0, 0),
  initial_collection_data = c(60, 60, 60), start = 10)
}
\arguments{
\item{y}{is the number of units used at the current day i}

\item{t_pred}{is the sum of the predicted number of units for days
i+1, i+2, i+3}

\item{initial_expiry_data}{the number of units that can be used the
current day and the next (2-length vector)}

\item{initial_collection_data}{is the number of units to collect
for days i, i+1, i+2}

\item{start}{is when we start the clock. So for example, in this
default invocation the shelf storage (initial_collection_data)
was 60, 60, 60, on days 10, 11, and 12. So the evaluation of
the model begins at day 11 but the decision to order collection
starts on day 13.}
}
\value{
a list with four components, \code{x} is the number of
    units to collect, \code{r} is a matrix of two columns
    indicating remaining units usable on day i+1 and day i+2,
    \code{w} is waste, and \code{s} is shortage.
}
\description{
Compute three-day predicted usage, waste, remaining and shortage
using the actual usage and predicted usage.
}
