% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_helper.R
\name{create_folds}
\alias{create_folds}
\title{Create folds for time series cross validation. Folds are continguous and sequential.}
\usage{
create_folds(n, fold_len)
}
\arguments{
\item{n}{number of rows of a matrix/data_frame to which ids are assigned}

\item{fold_len}{length of a fold}
}
\value{
a vector of integers assigning values in 1 - nfold to each row.
}
\description{
Create folds for time series cross validation. Folds are continguous and sequential.
}
