% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{cross_validate}
\alias{cross_validate}
\title{Run cross-validation method}
\usage{
cross_validate(
  data,
  pred_var_indices,
  resp_var_index,
  l1_bounds,
  lag_bounds,
  c0,
  start,
  penalty_factor,
  seed_prop,
  fold_size,
  cv_type = "loo",
  show_progress = FALSE
)
}
\arguments{
\item{data}{the full dataset (data.frame) used as input to the model}

\item{pred_var_indices}{the column indices corresponding to the features in "data"}

\item{resp_var_index}{the (single) column index corresponding to the response in "data}

\item{l1_bounds}{vector containing the possible values of the l1 bound on coefs aside from
days of the week and seven day moving average}

\item{lag_bounds}{vector containing possible values of the bound on the seven day moving average
(lag) coefficient}

\item{c0}{the lower bound on remaining "fresh" inventory (used in optimization)}

\item{start}{the first day in the dataset that the model's predictions are evaluated}

\item{penalty_factor}{how much we penalize shortage over waste}

\item{seed_prop}{the proportion of seed data (treated as fold 1 for training)}

\item{fold_size}{the desired size of each fold (aside from the seed fold 1)}

\item{cv_type}{the variety of cross_validation used. The model can either train on all previous
folds (exp for exanding) or on all other folds besides the left-out fold (loo for leave-one-out, default)}

\item{show_progress}{a TRUE/FALSE flag for showing progress, default FALSE}
}
\value{
a list containing matrices for waste, r1, r2, shortage, and predictions from cross-validation
}
\description{
Run cross-validation method
}
