% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_LP2.R
\name{cv_lpSolve}
\alias{cv_lpSolve}
\title{Solve an LP problem}
\usage{
cv_lpSolve(d, ind, lambds, start = 10, c = 30)
}
\arguments{
\item{d}{the data frame}

\item{ind}{representing the fold}

\item{lambds}{the lambdas to use in penalization}

\item{start}{the start date}

\item{c}{the value for c_0}
}
\description{
Solve an LP problem
}
