% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day3_data}
\alias{day3_data}
\title{Day 3 data for use in prediction}
\format{A data frame with 880 rows and 44 variables:
\describe{
  \item{day}{The day number}
  \item{M,T,W,Th,F,S,Sn}{Day of the week}
  \item{seven_lag}{}
  \item{PLTlow}{}
  \item{WBChigh}{}
  \item{RBClow}{}
  \item{HGBlow}{}
  \item{MCVlow}{}
  \item{MCHlow}{}
  \item{MCHClow}{}
  \item{RDW}{}
  \item{ALYM}{}
  \item{B1}{}
  \item{B2}{}
  \item{B3}{}
  \item{C1}{}
  \item{C2}{}
  \item{C3}{}
  \item{CAPR.XFER.OVERFL}{}
  \item{CATH.PACU}{}
  \item{CDU.CLIN.DEC.UNIT}{}
  \item{D1CC}{}
  \item{D1CS}{}
  \item{D2}{}
  \item{D3}{}
  \item{DGR}{}
  \item{E1}{}
  \item{E2.ICU}{}
  \item{E29.ICU}{}
  \item{E3}{}
  \item{EMERGENCY.DEPARTMENT}{}
  \item{F3}{}
  \item{FGR}{}
  \item{G1}{}
  \item{G2P}{}
  \item{G2S}{}
  \item{H1}{}
  \item{H2}{}
  \item{plt.used}{the amount of platelet used}
}}
\usage{
day3_data
}
\description{
Day 3 data for use in prediction
}
\references{
doi:10.1073/pnas.1714097114
}
\keyword{data}
