% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_helper.R
\name{ma_week}
\alias{ma_week}
\title{Moving average of value for specific day of week over last x weeks}
\usage{
ma_week(v, x = 6L)
}
\arguments{
\item{v}{a vector of values (assume they are ordered chronologically and contiguous)}

\item{x}{integer number of weeks}
}
\value{
a vector average of the values for the last x weeks, scaled appropriately to avoid NAs
}
\description{
Moving average of value for specific day of week over last x weeks
}
