% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip.R
\docType{package}
\name{pip-package}
\alias{pip-package}
\alias{platelet}
\title{pip: Platelet Inventory Prediction using Linear Programming}
\description{
Forecast inventory of platelet or any other item based on past
history and projected usage. Facilities are provided for indicating
forecast window and length of past history to use. A combination of
time series analysis together with linear programming is used to
optimize a criterion.
}
\author{
Leying Guan, Robert J. Tibshirani, Balasubramanian Narasimhan, Alison Zemek, Xiaoying Tian, Robert Scott, Saurabh Gombar, Tho Duc Pham, Gomathi Krishnan

Maintainer: Robert J. Tibshirani<tibs@stanford.edu>
}
\keyword{package}
