% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{build_model}
\alias{build_model}
\title{Run a fit on a given dataset}
\usage{
build_model(
  data,
  expiry,
  collection,
  c0 = 15,
  history_window = 200,
  penalty_factor = 15,
  start = 10,
  min_lambda = 0,
  date_column = "day|date",
  response_column = "plt_used",
  show_progress = TRUE
)
}
\arguments{
\item{data}{the dataset}

\item{expiry}{the number of units expiring in an day and the day after (each day),
a list of 2 vectors}

\item{collection}{is the number of units that will be collected each day of training}

\item{c0}{the c0 value}

\item{history_window}{Number of days to look back}

\item{penalty_factor}{penalty for shortage specified by doctors}

\item{start}{the day the model is evaluated??. Default 10}

\item{min_lambda}{the smallest acceptable value of lambda (cap on coefficient magnitude)}

\item{date_column}{the name of the date or day number column as a
regex, default is "day|date" i.e. day or date}

\item{response_column}{the name of the response column, default is
"plt_used"}

\item{show_progress}{a TRUE/FALSE flag for showing progress,
default TRUE}
}
\description{
The dataset should have one more column than the predictors and
response. This additional column, which is typically a date or day
number, has to be named "date" or "day"
}
\examples{
build_model(c0 = 30,
           history_window = 200,
           penalty_factor = 15,
           start = 10,
           response_column = "plt.used",
           data = day3_data,
           show_progress = FALSE)

}
