% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{compute_loss}
\alias{compute_loss}
\title{Compute the Cross-Validation Loss for each hyperparameter}
\usage{
compute_loss(
  preds,
  y,
  w,
  r1,
  r2,
  s,
  penalty_factor,
  lo_inv_limit,
  hi_inv_limit
)
}
\arguments{
\item{preds}{a vector of 3 day usage predictions}

\item{y}{a vector of actual daily usage}

\item{w}{a vector of waste determined by pip::compute_prediction_statistics}

\item{r1}{a vector of remaining inventory determined by pip::compute_prediction_statistics}

\item{r2}{a vector of remaining inventory determined by pip::compute_prediction_statistics}

\item{s}{a vector of inventory shortage as determined by pip::compute_prediction_statistics}

\item{penalty_factor}{factor to additionally penalize shortage terms over waste}

\item{lo_inv_limit}{threshold for insufficient inventory below which to penalize (shortage risk)}

\item{hi_inv_limit}{threshold for surfeit inventory above which to penalize (wastage risk)}
}
\description{
Compute the Cross-Validation Loss for each hyperparameter
}
