% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Run a fit on a given dataset}
\usage{
evaluate_model(
  data,
  c0 = 15,
  train_window = 200,
  test_window = 7,
  penalty_factor = 5,
  lo_inv_limit = 30,
  hi_inv_limit = 60,
  start = 10,
  l1_bounds = seq(from = 200, to = 0, by = -2),
  lag_bounds = c(NA),
  date_column = "day|date",
  response_column = "plt_used"
)
}
\arguments{
\item{data}{the dataset}

\item{c0}{the c0 value}

\item{train_window}{Number of days to look back}

\item{test_window}{Number of days on which to evaluate model}

\item{penalty_factor}{penalty for shortage specified by doctors}

\item{start}{the day you start evaluating the model (for CV)}

\item{l1_bounds}{vector containing the possible values of the l1 bound on coefs aside from
days of the week and seven day moving average}

\item{lag_bounds}{vector containing possible values of the bound on the seven day moving average
(lag) coefficient}

\item{date_column}{the name of the date or day number column as a
regex, default is "day|date" i.e. day or date}

\item{response_column}{the name of the response column, default is
"plt_used"}
}
\description{
The dataset should have one more column than the predictors and
response. This additional column, which is typically a date or day
number, has to be named "date" or "day"
}
\examples{

}
