% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{predict_three_day_sum}
\alias{predict_three_day_sum}
\title{Predict the next three day sum and units to order if available}
\usage{
predict_three_day_sum(model, new_data)
}
\arguments{
\item{model}{the trained model}

\item{new_data}{a new data frame with the same predictor names as in
the training data, even if one row}
}
\value{
the predicted three day total of how many units to collect
}
\description{
Predict the next three day sum and units to order if available
}
