% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_LP2_commented.R
\name{single_lpSolve}
\alias{single_lpSolve}
\title{Solve an LP problem}
\usage{
single_lpSolve(d, lamb, start = 10, c = 30, buffer = 10,
  ind = NULL)
}
\arguments{
\item{d}{the data frame}

\item{lamb}{the lambda value}

\item{start}{the start date}

\item{c}{the value for c_0}

\item{buffer}{a value to be added to account for training in the initial stages, default 10}

\item{ind}{the indices for cross validation used only if non-null. (Should be sorted!!)}
}
\description{
Solve an LP problem
}
