% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_LP2_commented.R
\name{single_lpSolve}
\alias{single_lpSolve}
\title{Solve an LP problem}
\usage{
single_lpSolve(
  d,
  l1_bounds,
  lag_bounds,
  num_vars,
  start = 10,
  c = 30,
  buffer = 10,
  ind = NULL
)
}
\arguments{
\item{d}{the data frame}

\item{l1_bounds}{the upper bound on the l1 norm of coefficient vector (aside from lag and days of week)}

\item{lag_bounds}{the upper bound on the seven day moving average of usage coefficient (lag)}

\item{num_vars}{the number of features/covariates in the dataset.}

\item{start}{the start date}

\item{c}{the value for c_0}

\item{buffer}{a value to be added to account for training in the initial stages, default 10}

\item{ind}{the indices for cross validation used only if non-null. (Should be sorted!!)}
}
\description{
Solve an LP problem
}
